+++ packages/ionic-react/src/schematics/application/lib/files.ts
@@ -21,12 +21,12 @@ export function addBaseTemplate(options: NormalizedSchema): Rule {
       applyTemplates({
         ...options,
         ...names(options.name),
-        offsetFromRoot: offsetFromRoot(options.projectRoot),
+        offsetFromRoot: offsetFromRoot(options.appProjectRoot),
       }),
       options.unitTestRunner === 'none'
         ? filter((file) => !file.endsWith('.spec.tsx'))
         : noop(),
-      move(options.projectRoot),
+      move(options.appProjectRoot),
       options.js ? toJS() : noop(),
     ]),
     MergeStrategy.Overwrite
@@ -39,12 +39,12 @@ export function addBlankTemplate(options: NormalizedSchema): Rule {
       applyTemplates({
         ...options,
         ...names(options.name),
-        offsetFromRoot: offsetFromRoot(options.projectRoot),
+        offsetFromRoot: offsetFromRoot(options.appProjectRoot),
       }),
       options.unitTestRunner === 'none'
         ? filter((file) => !file.endsWith('.spec.tsx'))
         : noop(),
-      move(options.projectRoot),
+      move(options.appProjectRoot),
       options.js ? toJS() : noop(),
     ]),
     MergeStrategy.Overwrite
@@ -57,8 +57,8 @@ export function addFiles(options: NormalizedSchema): Rule {
 
 export function deleteUnusedFiles(options: NormalizedSchema): Rule {
   return (tree: Tree) => {
-    tree.delete(options.projectRoot + '/src/favicon.ico');
-    tree.delete(options.projectRoot + `/src/app/App.css`);
+    tree.delete(options.appProjectRoot + '/src/favicon.ico');
+    tree.delete(options.appProjectRoot + `/src/app/App.css`);
 
     return tree;
   };