+++ packages/ionic-react/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Features
 
 - support custom Nx layouts
+- update Ionic to 5.5.1
 
 ## BREAKING CHANGES